# Conflicts:
#	TeachSync/src/main/resources/TeachSyncV5.sql

+++ TeachSync/src/main/java/com/teachsync/controllers/ClazzController.java
@@ -3,8 +3,10 @@ package com.teachsync.controllers;
 import com.teachsync.dtos.clazz.ClazzCreateDTO;
 import com.teachsync.dtos.clazz.ClazzReadDTO;
 import com.teachsync.dtos.clazz.ClazzUpdateDTO;
+import com.teachsync.dtos.user.UserReadDTO;
 import com.teachsync.services.clazz.ClazzService;
 import com.teachsync.services.course.CourseService;
+import com.teachsync.utils.Constants;
 import com.teachsync.utils.enums.Status;
 import jakarta.servlet.http.HttpServletRequest;
 import org.modelmapper.ModelMapper;
@@ -16,6 +18,7 @@ import org.springframework.util.ObjectUtils;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.SessionAttribute;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import java.util.Set;
@@ -30,7 +33,6 @@ public class ClazzController {
 
     @Autowired
     private ModelMapper mapper;
-
 
 
     @GetMapping("/clazz")
@@ -64,8 +66,20 @@ public class ClazzController {
     @PostMapping("/add-clazz")
     public String addClazz(
             HttpServletRequest request,
+            @SessionAttribute(value = "user", required = false) UserReadDTO userDTO,
             Model model,
             RedirectAttributes redirect) throws Exception {
+        //check login
+        if (ObjectUtils.isEmpty(userDTO)) {
+            redirect.addAttribute("mess", "Làm ơn đăng nhập");
+            return "redirect:/";
+        }
+
+        if (!userDTO.getRoleId().equals(Constants.ROLE_ADMIN)) {
+            redirect.addAttribute("mess", "bạn không đủ quyền");
+            return "redirect:/";
+        }
+
         ClazzCreateDTO createDTO = new ClazzCreateDTO();
         createDTO.setClazzName(request.getParameter("name"));
         createDTO.setClazzDesc(request.getParameter("desc"));
@@ -98,7 +112,21 @@ public class ClazzController {
     }
 
     @GetMapping("/delete-clazz")
-    public String deleteClazz(HttpServletRequest request, Model model, RedirectAttributes redirect) {
+    public String deleteClazz(
+            @SessionAttribute(value = "user", required = false) UserReadDTO userDTO,
+            HttpServletRequest request,
+            Model model,
+            RedirectAttributes redirect) {
+        //check login
+        if (ObjectUtils.isEmpty(userDTO)) {
+            redirect.addAttribute("mess", "Làm ơn đăng nhập");
+            return "redirect:/";
+        }
+
+        if (!userDTO.getRoleId().equals(Constants.ROLE_ADMIN)) {
+            redirect.addAttribute("mess", "bạn không đủ quyền");
+            return "redirect:/";
+        }
         Long Id = Long.parseLong(request.getParameter("Id"));
         String result = clazzService.deleteClazz(Id);
         if (result.equals("success")) {
